Add support for Python syntax checking
The CI server will now read JSON data sent with POST, clone a repository
and attempt to parse every Python file in the repository to detect
syntax errors. This also adds two unit tests, one that submits a
repository with valid code and one that submits a repository with
invalid code.

Fixes #4

Co-authored-by: editf <editf@users.noreply.github.com>