Make parser ignore test directory
This makes the try_compile_all method ignore any files in directories
called test so that files that are invalid on purpose do not cause the
syntax check to fail. To ensure that this behaviour remains, a new unit
test has been added. In order to give this unit test access to the
try_compile_all method, it has been converted into a public static
method.

Fixes #12