Implement setting commit status
This adds a function that lets the server set the commit status of a recently pushed commit.
It will be set to "failure" if there are syntax errors, "success" if there are no errors
and all tests suceed, and "error" if there are no tests or at least one test failed.

Fixes #6

Co-authored-by: Penguinification <Penguinification@users.noreply.github.com>
Co-authored-by: Ahmad627 <Ahmad627@users.noreply.github.com>
Co-authored-by: Erabl <Erabl@users.noreply.github.com>