Implement build history
This adds support for showing the build history of the CI server.
It can be accessed as an HTML document by sending a GET request to the
root directory of the server. Additionally, each build can be accessed
individually by sending a GET request to the server with the SHA as the
path, as expected from a RESTful API. Unit tests have also been added.

Fixes #28

Co-authored-by: editf <editf@users.noreply.github.com>